# Conflicts:
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/concordium/wallet/core/crypto/CryptoLibraryMock.kt
#	app/src/main/java/com/concordium/wallet/core/crypto/CryptoLibraryReal.kt
#	app/src/main/java/com/concordium/wallet/data/backend/repository/ProxyRepository.kt
#	app/src/main/java/com/concordium/wallet/data/cryptolib/CreateTransferInput.kt
#	app/src/main/java/com/concordium/wallet/data/model/Transaction.kt
#	app/src/main/java/com/concordium/wallet/data/room/Account.kt
#	app/src/main/java/com/concordium/wallet/data/room/WalletDatabase.kt
#	app/src/main/java/com/concordium/wallet/data/util/CurrencyUtil.kt
#	app/src/main/java/com/concordium/wallet/ui/account/accountdetails/AccountDetailsActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/account/accountdetails/AccountDetailsViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/account/accountdetails/AccountReleaseScheduleActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/account/accountsoverview/AccountsOverviewFragment.kt
#	app/src/main/java/com/concordium/wallet/ui/account/accountsoverview/AccountsOverviewViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/account/common/accountupdater/AccountUpdater.kt
#	app/src/main/java/com/concordium/wallet/ui/bakerdelegation/baker/BakerRegisterAmountActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/bakerdelegation/common/BaseDelegationBakerRegisterAmountActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/bakerdelegation/common/DelegationBakerViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/bakerdelegation/common/StakeAmountInputValidator.kt
#	app/src/main/java/com/concordium/wallet/ui/bakerdelegation/delegation/DelegationRegisterConfirmationActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/common/BackendErrorHandler.kt
#	app/src/main/java/com/concordium/wallet/ui/common/TransactionViewHelper.kt
#	app/src/main/java/com/concordium/wallet/ui/common/delegates/AuthDelegate.kt
#	app/src/main/java/com/concordium/wallet/ui/more/dev/DevViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/passphrase/recover/RecoverWalletActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/passphrase/recoverprocess/RecoverProcessViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/transaction/sendfunds/AddMemoActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/transaction/sendfunds/SendFundsActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/transaction/sendfunds/SendFundsViewModel.kt
#	app/src/main/java/com/concordium/wallet/ui/transaction/transactiondetails/TransactionDetailsActivity.kt
#	app/src/main/java/com/concordium/wallet/ui/walletconnect/WalletConnectViewModel.kt
#	app/src/main/java/com/concordium/wallet/uicore/ButtonsSlider.kt
#	app/src/main/java/com/concordium/wallet/uicore/MaxAmountTextWatcher.kt
#	app/src/main/java/com/concordium/wallet/uicore/view/AmountEditText.kt
#	app/src/main/res/drawable/ic_bg_shield.xml
#	app/src/main/res/layout/activity_account_details.xml
#	app/src/main/res/layout/activity_account_settings.xml
#	app/src/main/res/layout/activity_add_memo.xml
#	app/src/main/res/layout/activity_identity_provider_webview.xml
#	app/src/main/res/layout/activity_recover_process.xml
#	app/src/main/res/layout/activity_transaction_details.xml
#	app/src/main/res/layout/item_transaction.xml
#	app/src/main/res/values/strings.xml

+++ app/src/androidTest/java/com/concordium/wallet/data/room/TransferDaoUnitTest.kt
@@ -46,8 +46,8 @@ class TransferDaoUnitTest {
         val transfer = Transfer(
             1,
             1,
-            100,
-            20,
+            100.toBigInteger(),
+            20.toBigInteger(),
             "0",
             "0",
             1901176193,
@@ -90,8 +90,8 @@ class TransferDaoUnitTest {
         return Transfer(
             0,
             1,
-            100,
-            20,
+            100.toBigInteger(),
+            20.toBigInteger(),
             "0",
             "0",
             1901176193,
